Updated /Includes.xml with script.artistslideshow
Option to view current playing image of the artist. It has some bugs:
Only works when Control.HasFocus(502) (Don't know how to do it to make it visible always)
And also it doesnt start till you press tab on the keyboard. Once you pressed tab and go back to Home, the image appears and work great :)